Move environment to another file
