+++ j5/backends/__init__.py
@@ -1,5 +1,6 @@
 """Backend classes."""
 
-from .backend import Backend, BackendMeta, CommunicationError, Environment
+from .backend import Backend, BackendMeta, CommunicationError
+from .environment import Environment
 
 __all__ = ["Backend", "BackendMeta", "CommunicationError", "Environment"]
